Fix device
